pyproject.toml
[project]
name = "ccpa-compliance-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.134.0",
    "huggingface-hub>=1.5.0",
    "llama-cpp-python>=0.3.16",
    "pydantic>=2.12.5",
    "pypdf>=6.7.4",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "uvicorn>=0.41.0",
]
